requirements.txt
mysql-connector-python
neo4j
Pillow
pytz
pika
flask